requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
frozendict==2.4.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.6
lxml==5.2.2
multitasking==0.0.11
numpy==2.0.0
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
webencodings==0.5.1
yfinance==0.2.40
